created tests for all the user use cases, modified the files to reflect bugs found during testing

+++ src/use-cases/user/deleteUser.ts
@@ -1,4 +1,5 @@
 import InternalServerError from "../../errors/internalServerError.js";
+import ValidationError from "../../errors/validationError.js";
 import type IUseCase from "../../interfaces/useCase.js";
 import type IUserDAO from "../../interfaces/user/userDAO.js";
 
@@ -7,6 +8,9 @@ export default class DeleteUser implements IUseCase<boolean> {
         private userDAO: IUserDAO
     ) { }
     async call(userId: string) {
+        if (typeof userId !== "string"){
+            throw new ValidationError("userId must be a string")
+        }
         let deleted: boolean
         try {
             deleted = await this.userDAO.delete(userId)
